WEBUI-1987: Extract inactivity timer into NuxeoInactivityBehavior
Address review feedback: move the client-side inactivity timer, cross-tab
activity sharing, server keep-alive and 401 logout-redirect logic out of the
already-large nuxeo-app.js into a dedicated Polymer behavior
(elements/behaviors/nuxeo-inactivity-behavior.js), composed via the behaviors
array - mirroring the existing NuxeoAppDrawerResizeBehavior pattern and the
AGENTS.md convention for cross-cutting logic.

No behavior change: the methods are mixed into nuxeo-app and still wired from
its ready()/attached()/detached() lifecycle; _logout() stays on nuxeo-app
(template binding for the sign-out menu) and the keepAlive <nuxeo-resource>
stays in the template. All unit tests pass (2254).

(cherry picked from commit db2387fc77ca79e83e08a43bc45d9b2f5af7ae73)
